Major refactor:
personally, I feel like the following functions:

  goawait.Poll()
  goawait.PollFirst()
  goawait.PollAll()

aren't very clear and don't express the intent of the functions.

i'd just name the package "await" and call those functions:

  await.Func() - wait for one function to complete
  await.First() - wait for the first of a function map to complete
  await.All() - wait for all of the functions in a function map to
  complete.

this seems like a better naming scheme that has a simple verb-noun
structure that consists of real words that make the api easier to read.

this change required changing the signatures of all of the functions and
updating the package directives in each source file.